Functioning site with protections from getting anywhere without being logged in
I went through the site for a few hours just messing around with css.  I found a couple of errors in the authentication, but idk if it may have been from merge conflicts that followed into this new repo?  I changed everywhere we look for session[:id] to where it's a call to our helper method is_authenticated?  I made a view to put our error methods if we want to display what happens when trying to access decks page without being logged in.